Refactor and add simple README and LICENSE
